Move LocationClientProvider to location module
