+++ location/src/main/java/org/odk/collect/location/LocationClient.java
@@ -2,11 +2,11 @@ package org.odk.collect.location;
 
 import android.location.Location;
 
-import com.google.android.gms.location.LocationListener;
-import com.google.android.gms.location.LocationRequest;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+
+import com.google.android.gms.location.LocationListener;
+import com.google.android.gms.location.LocationRequest;
 
 /**
  * An interface for classes that allow monitoring and retrieving the User's Location.
